Add more tests for parameterized tests
