+++ test/E2ETests/DiscoveryAndExecutionTests/Parameterized tests/DynamicDataTests.cs
@@ -1,0 +1,70 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.MSTestV2.CLIAutomation;
+
+namespace Microsoft.MSTestV2.Smoke.DiscoveryAndExecutionTests;
+public class DynamicDataTests : CLITestBase
+{
+    private const string TestAssembly = "DynamicDataTestProject.dll";
+
+    public void ExecuteDynamicDataTests()
+    {
+        // Arrange
+        var assemblyPath = GetAssetFullPath(TestAssembly);
+
+        // Act
+        var testCases = DiscoverTests(assemblyPath);
+        var testResults = RunTests(assemblyPath, testCases);
+
+        // Assert
+        VerifyE2E.ContainsTestsPassed(
+            testResults,
+            "DynamicDataTest_SourceMethod (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourceMethod (Jane;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourceProperty (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourceProperty (Jane;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceMethod_CustomDisplayName with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceMethod_CustomDisplayName with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceProperty_CustomDisplayName with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceProperty_CustomDisplayName with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceMethod_CustomDisplayNameOtherType with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceMethod_CustomDisplayNameOtherType with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceProperty_CustomDisplayNameOtherType with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceProperty_CustomDisplayNameOtherType with 2 parameters",
+            "DynamicDataTest_SourceMethodOtherType (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourceMethodOtherType (Jane;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourcePropertyOtherType (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTest_SourcePropertyOtherType (Jane;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceMethodOtherType_CustomDisplayName with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourceMethodOtherType_CustomDisplayName with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceMethodOtherType_CustomDisplayNameOtherType with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourceMethodOtherType_CustomDisplayNameOtherType with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourcePropertyOtherType_CustomDisplayName with 2 parameters",
+            "Custom DynamicDataTestMethod DynamicDataTest_SourcePropertyOtherType_CustomDisplayName with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourcePropertyOtherType_CustomDisplayNameOtherType with 2 parameters",
+            "UserDynamicDataTestMethod DynamicDataTest_SourcePropertyOtherType_CustomDisplayNameOtherType with 2 parameters",
+            "DynamicDataTestWithTestCategory (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTestWithTestCategory (Jane;Doe,LibProjectReferencedByDataSourceTest.User)");
+
+        VerifyE2E.FailedTestCount(testResults, 0);
+    }
+
+    public void ExecuteDynamicDataTestsWithCategoryFilter()
+    {
+        // Arrange
+        var assemblyPath = GetAssetFullPath(TestAssembly);
+
+        // Act
+        var testCases = DiscoverTests(assemblyPath, "TestCategory~DynamicDataWithCategory");
+        var testResults = RunTests(assemblyPath, testCases);
+
+        // Assert
+        VerifyE2E.ContainsTestsPassed(
+            testResults,
+            "DynamicDataTestWithTestCategory (John;Doe,LibProjectReferencedByDataSourceTest.User)",
+            "DynamicDataTestWithTestCategory (Jane;Doe,LibProjectReferencedByDataSourceTest.User)");
+
+        VerifyE2E.FailedTestCount(testResults, 0);
+    }
+}
