+++ test/E2ETests/Smoke.E2E.Tests/Parameterized tests/DataRowTests.cs
@@ -4,22 +4,91 @@
 using Microsoft.MSTestV2.CLIAutomation;
 
 namespace MSTestAdapter.Smoke.E2ETests;
-public class TestDataSourceTests : CLITestBase
+public class DataRowTests : CLITestBase
 {
-    private const string TestAssembly = "DataSourceTestProject.dll";
+    private const string TestAssembly = "DataRowTestProject.dll";
 
-    public void ExecuteDynamicDataTests()
+    public void ExecuteOnlyDerivedClassDataRowsWhenBothBaseAndDerivedClassHasDataRows_SimpleDataRows()
     {
-        // Arrange & Act
-        InvokeVsTestForExecution(
-            new string[] { TestAssembly },
-            testCaseFilter: "DynamicDataTest");
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "TestCategory~DataRowSimple");
 
-        // Assert
-        ValidatePassedTests(
-            "DynamicDataTest (John;Doe,DataSourceTestProject.ITestDataSourceTests.User)");
+        ValidatePassedTestsContain(
+            "DataRowTestMethod (BaseString1)",
+            "DataRowTestMethod (BaseString2)",
+            "DataRowTestMethod (BaseString3)",
+            "DataRowTestMethod (DerivedString1)",
+            "DataRowTestMethod (DerivedString2)");
 
-        ValidateFailedTestsCount(0);
+        // 3 tests of BaseClass.DataRowTestMethod - 3 data row results and no parent result
+        // 2 tests of DerivedClass.DataRowTestMethod - 2 data row results and no parent result
+        // Total 5 tests - Making sure that DerivedClass doesn't run BaseClass tests
+        ValidatePassedTestsCount(5);
+    }
+
+    public void ExecuteOnlyDerivedClassDataRowsWhenItOverridesBaseClassDataRows_SimpleDataRows()
+    {
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "FullyQualifiedName~DerivedClass&TestCategory~DataRowSimple");
+
+        ValidatePassedTestsContain(
+            "DataRowTestMethod (DerivedString1)",
+            "DataRowTestMethod (DerivedString2)");
+
+        // 2 tests of DerivedClass.DataRowTestMethod - 2 datarow result and no parent result
+        ValidatePassedTestsCount(2);
+    }
+
+    public void DataRowsExecuteWithRequiredAndOptionalParameters()
+    {
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "TestCategory~DataRowSomeOptional");
+
+        ValidatePassedTestsContain(
+            "DataRowTestMethodWithSomeOptionalParameters (123)",
+            "DataRowTestMethodWithSomeOptionalParameters (123,DerivedOptionalString1)",
+            "DataRowTestMethodWithSomeOptionalParameters (123,DerivedOptionalString2,DerivedOptionalString3)");
+
+        // 3 tests of DerivedClass.DataRowTestMethodWithSomeOptionalParameters - 3 datarow result and no parent result
+        ValidatePassedTestsCount(3);
+    }
+
+    public void DataRowsExecuteWithAllOptionalParameters()
+    {
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "TestCategory~DataRowAllOptional");
+
+        ValidatePassedTestsContain(
+            "DataRowTestMethodWithAllOptionalParameters ()",
+            "DataRowTestMethodWithAllOptionalParameters (123)",
+            "DataRowTestMethodWithAllOptionalParameters (123,DerivedOptionalString4)",
+            "DataRowTestMethodWithAllOptionalParameters (123,DerivedOptionalString5,DerivedOptionalString6)");
+
+        // 4 tests of DerivedClass.DataRowTestMethodWithAllOptionalParameters - 4 datarow result and no parent result
+        ValidatePassedTestsCount(4);
+    }
+
+    public void DataRowsExecuteWithParamsArrayParameter()
+    {
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "TestCategory~DataRowParamsArgument");
+
+        ValidatePassedTestsContain(
+            "DataRowTestMethodWithParamsParameters (2)",
+            "DataRowTestMethodWithParamsParameters (2,DerivedSingleParamsArg)",
+            "DataRowTestMethodWithParamsParameters (2,DerivedParamsArg1,DerivedParamsArg2)",
+            "DataRowTestMethodWithParamsParameters (2,DerivedParamsArg1,DerivedParamsArg2,DerivedParamsArg3)");
+
+        // 4 tests of DerivedClass.DataRowTestMethodWithParamsParameters - 4 datarow result and no parent result
+        ValidatePassedTestsCount(4);
+    }
+
+    public void DataRowsFailWhenInvalidArgumentsProvided()
+    {
+        InvokeVsTestForExecution(new string[] { TestAssembly }, testCaseFilter: "FullyQualifiedName~DataRowTests_Regular&TestCategory~DataRowOptionalInvalidArguments");
+
+        ValidatePassedTestsContain(
+            "DataRowTestMethodFailsWithInvalidArguments ()",
+            "DataRowTestMethodFailsWithInvalidArguments (2)",
+            "DataRowTestMethodFailsWithInvalidArguments (2,DerivedRequiredArgument,DerivedOptionalArgument,DerivedExtraArgument)");
+
+        // 3 tests of DerivedClass.DataRowTestMethodFailsWithInvalidArguments - 3 datarow result and no parent result
+        ValidatePassedTestsCount(3);
     }
 
     public void ExecuteDataRowTests_Enums()
@@ -105,7 +174,7 @@ public class TestDataSourceTests : CLITestBase
         ValidatePassedTests(
             "DataRowNonSerializable (System.String)",
             "DataRowNonSerializable (System.Int32)",
-            "DataRowNonSerializable (DataSourceTestProject.ITestDataSourceTests.DataRowTests_Enums)");
+            "DataRowNonSerializable (DataRowTestProject.DataRowTests_Enums)");
 
         ValidateFailedTestsCount(0);
     }
@@ -115,7 +184,7 @@ public class TestDataSourceTests : CLITestBase
         // Arrange & Act
         InvokeVsTestForExecution(
             new string[] { TestAssembly },
-            testCaseFilter: "FullyQualifiedName~Regular_DataRowTests");
+            testCaseFilter: "FullyQualifiedName~DataRowTests_Regular");
 
         // Assert
         ValidatePassedTests(
@@ -140,24 +209,5 @@ public class TestDataSourceTests : CLITestBase
             "NullValueInData (john.doe@example.com,abc123,/unit/test)");
 
         ValidateFailedTestsCount(0);
-    }
-
-    // TODO @haplois | @evangelink: This test fails under CI - will be fixed in a future PR (marked as private to ignore the test)
-    private void ExecuteCsvTestDataSourceTests()
-    {
-        // Arrange & Act
-        InvokeVsTestForExecution(
-            new string[] { TestAssembly },
-            testCaseFilter: "CsvTestMethod");
-
-        // Assert
-        ValidatePassedTests(
-            "CsvTestMethod (Data Row 0)",
-            "CsvTestMethod (Data Row 2)");
-
-        ValidateFailedTests(
-            TestAssembly,
-            "CsvTestMethod (Data Row 1)",
-            "CsvTestMethod (Data Row 3)");
     }
 }
