+++ test/E2ETests/TestAssets/DataRowTestProject/DataRowTests_BaseClass.cs
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information. 
 
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace DataRowTestProject;
+
 [TestClass]
-public class BaseClass
+public class DataRowTests_BaseClass
 {
     [TestCategory("DataRowSimple")]
     [TestMethod]
