+++ test/E2ETests/TestAssets/DataRowTestProject/DataRowTests_Enums.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-using System;
+namespace DataRowTestProject;
 
-namespace DataSourceTestProject.ITestDataSourceTests;
 [TestClass]
 public class DataRowTests_Enums
 {
