+++ test/E2ETests/TestAssets/DataRowTestProject/DataRowTests_NonSerializablePaths.cs
@@ -1,12 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
 
-namespace DataSourceTestProject.ITestDataSourceTests;
+namespace DataRowTestProject;
+
 [TestClass]
 public class DataRowTests_NonSerializablePaths
 {
