+++ test/E2ETests/TestAssets/DataRowTestProject/DataRowTests_Regular.cs
@@ -1,14 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-using System;
+namespace DataRowTestProject;
 
-namespace DataSourceTestProject.ITestDataSourceTests;
 [TestClass]
-public class Regular_DataRowTests
+public class DataRowTests_Regular
 {
     [TestMethod]
     [DataRow(10)]
