+++ test/E2ETests/TestAssets/DynamicDataTestProject/DynamicDataTestProject.csproj
@@ -1,4 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
     <TargetFramework>net462</TargetFramework>
     <IsPackable>false</IsPackable>
@@ -8,16 +9,7 @@
 
   <ItemGroup>
     <ProjectReference Include="$(RepoRoot)src\Adapter\MSTest.TestAdapter\MSTest.TestAdapter.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Update="a.csv">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
+    <ProjectReference Include="..\LibProjectReferencedByDataSourceTest\LibProjectReferencedByDataSourceTest.csproj" />
   </ItemGroup>
 
 </Project>
